initial small set of testbed APIs in IDL format
Just want to try out this format and see if we should have a full set of the same APIs in IDL format. The IDL format is a valid format and has same capabilities. But having the testbed APIs also available as IDL format would ensure compatibility. So the outcome should ideally not diff at all fro the YAML versions.